Seperated NewProjectModal into its own component
